+++ app/queries/users/queries/attributes.rb
@@ -8,22 +8,33 @@ class Users::Queries::Attributes < Query::Base
   after_initialize :set_options
 
   def execute
-    user ? prepare_attrs_by_user(user) : users.map { |u| prepare_attrs_by_user(u) }
+    user ? attrs_by_user(user) : users.map { |u| attrs_by_user(u) }
   end
 
   private
 
-  def prepare_attrs_by_user(user)
+  def attrs_by_user(user)
     attrs.each_with_object({}) do |attr, memo|
-      memo[attr] = user.respond_to?(attr) ?
-        user.send(attr) : Classifier.as_klass(self.class.name, attr).new(user).value
+      memo[attr] = user.respond_to?(attr) ? user.send(attr) : external_attribute(user, attr)
     end
+  end
+
+  def external_attribute(user, attr)
+    klass = Classifier.as_klass(self.class.name, attr)
+    users && klass.collection_supported? ?
+      external_attribute_from_collection(user, attr, klass) : klass.new(user).value
+  end
+
+  def external_attribute_from_collection(user, attr, klass)
+    @cache[attr] = klass.new(users, as_collection: true).value unless @cache[attr]
+    @cache[attr][user.mkey]
   end
 
   def set_options
     @user  = User.find_by(mkey: options[:user]) if options[:user]
     @users = User.where(mkey: options[:users]) if options[:users]
     @attrs = Array.wrap(options[:attrs])
+    @cache = {}
   end
 
   #